refactor: move permissions out of util pkg
