+++ pkg/azure/client/application/application.go
@@ -11,11 +11,11 @@ import (
 
 	"github.com/nais/azureator/pkg/azure"
 	"github.com/nais/azureator/pkg/azure/client/application/approle"
-	groupmembershipclaim2 "github.com/nais/azureator/pkg/azure/client/application/groupmembershipclaim"
+	"github.com/nais/azureator/pkg/azure/client/application/groupmembershipclaim"
 	"github.com/nais/azureator/pkg/azure/client/application/permissionscope"
+	"github.com/nais/azureator/pkg/azure/permissions"
 	"github.com/nais/azureator/pkg/azure/transaction"
 	"github.com/nais/azureator/pkg/azure/util"
-	"github.com/nais/azureator/pkg/azure/util/permissions"
 )
 
 // Application tags
@@ -80,7 +80,7 @@ func (a Application) Register(tx transaction.Transaction) (*msgraph.Application,
 
 	req := util.Application(a.defaultTemplate(tx.Instance)).
 		ResourceAccess(access).
-		GroupMembershipClaims(groupmembershipclaim2.GroupMembershipClaimApplicationGroup).
+		GroupMembershipClaims(groupmembershipclaim.GroupMembershipClaimApplicationGroup).
 		AppRoles(roles.GetResult()).
 		RedirectUris(util.GetReplyUrlsStringSlice(tx.Instance)).
 		PermissionScopes(scopes.GetResult()).
@@ -123,7 +123,7 @@ func (a Application) Update(tx transaction.Transaction) (*msgraph.Application, e
 
 	// todo: remove 'groupClaimsIsDefined' predicate after grace period
 	if a.Config().Features.GroupsAssignment.Enabled && groupClaimsIsDefined {
-		builder.GroupMembershipClaims(groupmembershipclaim2.GroupMembershipClaimApplicationGroup)
+		builder.GroupMembershipClaims(groupmembershipclaim.GroupMembershipClaimApplicationGroup)
 	}
 
 	app := builder.Build()
