+++ pkg/azure/client/application/owners.go
@@ -6,7 +6,7 @@ import (
 	msgraph "github.com/nais/msgraph.go/v1.0"
 
 	"github.com/nais/azureator/pkg/azure"
-	directoryobject2 "github.com/nais/azureator/pkg/azure/directoryobject"
+	"github.com/nais/azureator/pkg/azure/directoryobject"
 	"github.com/nais/azureator/pkg/azure/transaction"
 )
 
@@ -24,12 +24,12 @@ func (o owners) Process(tx transaction.Transaction, desired []msgraph.DirectoryO
 		return err
 	}
 
-	newOwners := directoryobject2.Difference(desired, existing)
+	newOwners := directoryobject.Difference(desired, existing)
 	if err := o.registerFor(tx, newOwners); err != nil {
 		return fmt.Errorf("registering owners for application: %w", err)
 	}
 
-	revoked := directoryobject2.Difference(existing, desired)
+	revoked := directoryobject.Difference(existing, desired)
 	if err := o.revokeFor(tx, revoked); err != nil {
 		return fmt.Errorf("revoking owners for application: %w", err)
 	}
@@ -51,7 +51,7 @@ func (o owners) registerFor(tx transaction.Transaction, owners []msgraph.Directo
 	objectId := tx.Instance.GetObjectId()
 
 	for _, owner := range owners {
-		body := directoryobject2.ToOwnerPayload(owner)
+		body := directoryobject.ToOwnerPayload(owner)
 		req := o.GraphClient().Applications().ID(objectId).Owners().Request()
 		err := req.JSONRequest(tx.Ctx, "POST", "/$ref", body, nil)
 		if err != nil {
