+++ pkg/azure/client/application/permissionscope/permissionscope.go
@@ -5,7 +5,7 @@ import (
 	"github.com/nais/msgraph.go/ptr"
 	msgraph "github.com/nais/msgraph.go/v1.0"
 
-	"github.com/nais/azureator/pkg/azure/util/permissions"
+	"github.com/nais/azureator/pkg/azure/permissions"
 )
 
 func EnsureScopesRequireAdminConsent(scopes []msgraph.PermissionScope) []msgraph.PermissionScope {
