+++ pkg/azure/client/application/permissionscope/permissionscopemap.go
@@ -4,7 +4,7 @@ import (
 	"github.com/nais/msgraph.go/ptr"
 	msgraph "github.com/nais/msgraph.go/v1.0"
 
-	"github.com/nais/azureator/pkg/azure/util/permissions"
+	"github.com/nais/azureator/pkg/azure/permissions"
 )
 
 type Map map[string]msgraph.PermissionScope
